Sam's 3rd commit
Added methods HorizontalCheck, VerticalCheck, DiagonalsCheck to finish
isGameOver method (which determines the end of the game). Also renamed
whosNext method to whosUp.  Also added methods findXcord and findYcord
to finish the move method. These methods find the x and y coordinate of
the circle object created in move() which is dependent on row and column